pyproject.toml
[project]
name = "caso-aeo"
version = "0.1.0"
description = ""
authors = [
    {name = "Bruno Farfán",email = "bruno.farfan.miquel@uc.cl"}
]
requires-python = ">=3.13,<4.0"
dependencies = [
    "ruff (>=0.12.4,<0.13.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "openai (>=1.97.0,<2.0.0)",
    "pocketbase (>=0.15.0,<0.16.0)",
    "python-decouple (>=3.8,<4.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "duckduckgo-search (>=8.1.1,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
